+++ wrappers/java/kendo-taglib/src/main/java/com/kendoui/taglib/DateTimePickerTag.java
@@ -86,11 +86,11 @@ public class DateTimePickerTag extends WidgetTag /* interfaces */implements Anim
         setProperty("format", value);
     }
 
-    public int getInterval() {
-        return (int)getProperty("interval");
+    public float getInterval() {
+        return (float)getProperty("interval");
     }
 
-    public void setInterval(int value) {
+    public void setInterval(float value) {
         setProperty("interval", value);
     }
