+++ wrappers/java/kendo-taglib/src/main/java/com/kendoui/taglib/chart/LabelsTag.java
@@ -77,11 +77,11 @@ public class LabelsTag extends BaseTag /* interfaces */implements Border/* inter
         setProperty("format", value);
     }
 
-    public int getMargin() {
-        return (int)getProperty("margin");
+    public float getMargin() {
+        return (float)getProperty("margin");
     }
 
-    public void setMargin(int value) {
+    public void setMargin(float value) {
         setProperty("margin", value);
     }
 
@@ -93,35 +93,35 @@ public class LabelsTag extends BaseTag /* interfaces */implements Border/* inter
         setProperty("mirror", value);
     }
 
-    public int getPadding() {
-        return (int)getProperty("padding");
+    public float getPadding() {
+        return (float)getProperty("padding");
     }
 
-    public void setPadding(int value) {
+    public void setPadding(float value) {
         setProperty("padding", value);
     }
 
-    public int getRotation() {
-        return (int)getProperty("rotation");
+    public float getRotation() {
+        return (float)getProperty("rotation");
     }
 
-    public void setRotation(int value) {
+    public void setRotation(float value) {
         setProperty("rotation", value);
     }
 
-    public int getSkip() {
-        return (int)getProperty("skip");
+    public float getSkip() {
+        return (float)getProperty("skip");
     }
 
-    public void setSkip(int value) {
+    public void setSkip(float value) {
         setProperty("skip", value);
     }
 
-    public int getStep() {
-        return (int)getProperty("step");
+    public float getStep() {
+        return (float)getProperty("step");
     }
 
-    public void setStep(int value) {
+    public void setStep(float value) {
         setProperty("step", value);
     }
