+++ wrappers/java/kendo-taglib/src/main/java/com/kendoui/taglib/chart/LineTag.java
@@ -64,11 +64,11 @@ public class LineTag extends BaseTag /* interfaces *//* interfaces */ {
         setProperty("visible", value);
     }
 
-    public int getWidth() {
-        return (int)getProperty("width");
+    public float getWidth() {
+        return (float)getProperty("width");
     }
 
-    public void setWidth(int value) {
+    public void setWidth(float value) {
         setProperty("width", value);
     }
