+++ wrappers/java/kendo-taglib/src/main/java/com/kendoui/taglib/chart/PlotAreaTag.java
@@ -53,19 +53,19 @@ public class PlotAreaTag extends BaseTag /* interfaces */implements Border/* int
         setProperty("background", value);
     }
 
-    public int getOpacity() {
-        return (int)getProperty("opacity");
+    public float getOpacity() {
+        return (float)getProperty("opacity");
     }
 
-    public void setOpacity(int value) {
+    public void setOpacity(float value) {
         setProperty("opacity", value);
     }
 
-    public int getMargin() {
-        return (int)getProperty("margin");
+    public float getMargin() {
+        return (float)getProperty("margin");
     }
 
-    public void setMargin(int value) {
+    public void setMargin(float value) {
         setProperty("margin", value);
     }
