+++ wrappers/java/kendo-taglib/src/main/java/com/kendoui/taglib/grid/PageableTag.java
@@ -45,11 +45,11 @@ public class PageableTag extends BaseTag /* interfaces */implements Messages/* i
         setProperty("messages", value.properties());
     }
 
-    public int getPageSize() {
-        return (int)getProperty("pageSize");
+    public float getPageSize() {
+        return (float)getProperty("pageSize");
     }
 
-    public void setPageSize(int value) {
+    public void setPageSize(float value) {
         setProperty("pageSize", value);
     }
 
@@ -69,11 +69,11 @@ public class PageableTag extends BaseTag /* interfaces */implements Messages/* i
         setProperty("numeric", value);
     }
 
-    public int getButtonCount() {
-        return (int)getProperty("buttonCount");
+    public float getButtonCount() {
+        return (float)getProperty("buttonCount");
     }
 
-    public void setButtonCount(int value) {
+    public void setButtonCount(float value) {
         setProperty("buttonCount", value);
     }
