+++ wrappers/java/kendo-taglib/src/main/java/com/kendoui/taglib/lineargauge/MinorTicksTag.java
@@ -48,11 +48,11 @@ public class MinorTicksTag extends BaseTag /* interfaces *//* interfaces */ {
         setProperty("color", value);
     }
 
-    public int getSize() {
-        return (int)getProperty("size");
+    public float getSize() {
+        return (float)getProperty("size");
     }
 
-    public void setSize(int value) {
+    public void setSize(float value) {
         setProperty("size", value);
     }
 
@@ -64,11 +64,11 @@ public class MinorTicksTag extends BaseTag /* interfaces *//* interfaces */ {
         setProperty("visible", value);
     }
 
-    public int getWidth() {
-        return (int)getProperty("width");
+    public float getWidth() {
+        return (float)getProperty("width");
     }
 
-    public void setWidth(int value) {
+    public void setWidth(float value) {
         setProperty("width", value);
     }
